feat: 新增管理员概念 entity时间类型调整 #3111

+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/StrategyService.kt
@@ -123,7 +123,7 @@ class StrategyService @Autowired constructor(
             name = record.strategyName,
             strategy = JsonUtil.to(record.strategyBody),
             createUser = record.createUser,
-            createTime = DateTimeUtil.toDateTime(record.createTime).toLong()
+            createTime = DateTimeUtil.toDateTime(record.createTime)
         )
     }
